+++ homework/woo-junho/backend/db/users.js
@@ -1,0 +1,9 @@
+export default [
+    {
+        id: 1,
+        email: "noggong@example.com",
+        password: "1234",
+        name: "우준호",
+        role: "tutor",
+    }
+]
