+++ homework/woo-junho/backend/middlewares/leaveLog.middleware.js
@@ -1,0 +1,4 @@
+export default (req, res, next) => {
+    console.log(`${req.method} ${req.url} [${ new Date().toISOString()}] ${ req.headers.referer }`)
+    next()
+}
