+++ homework/woo-junho/backend/package.json
@@ -3,5 +3,6 @@
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2"
-  }
+  },
+  "type": "module"
 }
