+++ lesson/10_refactoring/esm/app.js
@@ -1,0 +1,6 @@
+import { sum, multiple, subtract } from './cal.js'
+// import calculator from './cal.js'
+
+console.log(sum(1, 2))
+console.log(multiple(1, 2))
+console.log(subtract(1, 2))
