+++ lesson/10_refactoring/esm/app2.js
@@ -1,0 +1,9 @@
+// import * as unknown from './cal.js'
+import * as cal  from './cal.js'
+
+console.log(cal)
+
+console.log(cal.sum(1, 2))
+console.log(cal.multiple(1, 2))
+console.log(cal.division(1, 2))
+console.log(cal.subtract(1, 2))
