+++ lesson/10_refactoring/esm/cal.js
@@ -1,0 +1,15 @@
+export function sum(a, b) {
+    return a + b
+}
+
+export function multiple(a, b) {
+    return a * b
+}
+export function division(a, b) {
+    return a / b
+}
+
+export function subtract(a, b) {
+    return a - b
+}
+
